requirements.txt
fastapi
uvicorn
pdf2image
Pillow
python-multipart
streamlit